+++ lib/mobility_active_storage/many_backend.rb
@@ -17,7 +17,7 @@ module MobilityActiveStorage
     include Mobility::Backend
     include BackendMethods
 
-    setup do |attributes, options|
+    setup do |attributes, options, backend_class|
       attributes.each do |attribute|
         options[:locales].each do |locale|
           has_many_attached :"#{attribute}_#{Mobility.normalize_locale(locale)}",
@@ -25,7 +25,8 @@ module MobilityActiveStorage
         end
 
         scope :"with_attached_#{attribute}", lambda {
-          includes("#{attribute}_#{Mobility.normalize_locale}_attachments": :blob)
+          name = backend_class.attachment_name_for(attribute, Mobility.locale)
+          includes("#{name}_attachments": :blob)
         }
 
         define_method(:"#{attribute}_locales") { mobility_backends[attribute].locales }